# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Ecommerce Backend API
  version: 1.0.0
  description: |
    OpenAPI specification for the API Gateway, BFF (Backend-for-Frontend), and all backend microservices of the ecommerce platform.

    ## Architecture Overview
    The system is composed of the following layers:
    - **API Gateway** (`:8080`) — Single ingress point. Validates JWT access tokens from the `access_token` cookie, enforces role-based access control, and proxies requests to the appropriate downstream service.
    - **BFF Service** (`:8088`) — Aggregates data from multiple services into single, page-optimised responses for the web frontend. All BFF endpoints also flow through the Gateway.
    - **Auth Service** (`:8081`) — Issues and validates JWT access tokens (short-lived) and refresh tokens (long-lived). Both tokens are delivered as HttpOnly cookies.
    - **User Service** (`:8085`) — Manages user profiles and credential changes.
    - **Product Service** (`:8082`) — Product catalogue, category tree, bulk import/export, and presigned S3 upload URLs.
    - **Cart Service** (`:8086`) — Per-user shopping cart and checkout initiation.
    - **Order Service** (`:8083`) — Order lifecycle management.
    - **Payment Service** (`:8087`) — Stripe payment session creation, verification, and webhook ingestion.
    - **Inventory Service** — Stock management, reservation, release, and confirmation as part of the order saga.
...

    ## Authentication
...

    ## Role-Based Access Control
...

...

...

...

...
servers:
...
paths:
...
  /bff/admin/users:
    get:
      tags: [Admin]
      summary: List all users (admin)
      operationId: adminListUsers
      description: |
        Returns a paginated list of all registered users (customers and admins).
        Passwords are never included in the response. Results are sorted by
        creation date descending (newest first). Requires admin role.
      parameters:
        - $ref: '#/components/parameters/AdminAuthorizationHeader'
        - $ref: '#/components/parameters/AdminUserIDHeader'
        - $ref: '#/components/parameters/AdminUserRoleHeader'
        - $ref: '#/components/parameters/AdminPageParam'
        - $ref: '#/components/parameters/AdminPageSizeParam'
      responses:
        '200':
          description: Paginated user list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      users:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              format: uuid
                            email:
                              type: string
                              format: email
                            name:
                              type: string
                            role:
                              type: string
                              enum: [user, admin]
                            phone_number:
                              type: string
                              nullable: true
                            created_at:
                              type: string
                              format: date-time
                      meta:
                        type: object
                        properties:
                          page:
                            type: integer
                          page_size:
                            type: integer
                          total:
                            type: integer
                          total_pages:
                            type: integer
                  error:
                    type: string
              example:
                success: true
                data:
                  users:
                    - id: e3bc8bbb-aee3-445c-b51a-0b3edee3b613
                      email: admin@example.com
                      name: Admin User
                      role: admin
                      created_at: '2026-03-10T01:46:38Z'
                  meta:
                    page: 1
                    page_size: 20
                    total: 23
                    total_pages: 2
                error: ''
        '403': {$ref: '#/components/responses/AdminForbiddenResponse'}
        '500': {$ref: '#/components/responses/AdminInternalErrorResponse'}
    post:
      tags: [Admin]
      summary: Create user (admin)
      operationId: adminCreateUser
      description: |
        Creates a user through the auth-service admin flow.
        Requires admin role and supports creating either `user` or `admin` accounts.
      parameters:
        - $ref: '#/components/parameters/AdminAuthorizationHeader'
        - $ref: '#/components/parameters/AdminUserIDHeader'
        - $ref: '#/components/parameters/AdminUserRoleHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password, role]
              properties:
                name:
                  type: string
                  minLength: 1
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                role:
                  type: string
                  enum: [admin, user]
            example:
              name: Ops Admin
              email: ops-admin@example.com
              password: StrongPass#123
              role: admin
      responses:
        '200': {$ref: '#/components/responses/AdminSuccessResponse'}
        '400': {$ref: '#/components/responses/AdminBadRequestResponse'}
        '403': {$ref: '#/components/responses/AdminForbiddenResponse'}
        '404': {$ref: '#/components/responses/AdminNotFoundResponse'}
        '500': {$ref: '#/components/responses/AdminInternalErrorResponse'}
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: |
        JWT access token stored as an HttpOnly cookie. Issued by the Auth Service on login or
        token refresh. Short-lived (typically 15 minutes). Pass the `admin` scope value in the
        `security` declaration to indicate an endpoint requires the `admin` role.
  parameters:
...
    AdminAuthorizationHeader:
      name: Authorization
      in: header
      required: true
      description: Bearer token header forwarded by gateway.
      schema:
        type: string
    AdminUserIDHeader:
      name: X-User-ID
      in: header
      required: true
      description: User identifier forwarded by API gateway.
      schema:
        type: string
    AdminUserRoleHeader:
      name: X-User-Role
      in: header
      required: true
      description: Role header forwarded by API gateway; must be `admin`.
      schema:
        type: string
        enum: [admin]
    AdminPageParam:
      name: page
      in: query
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    AdminPageSizeParam:
      name: page_size
      in: query
      description: Number of records per page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
...
    AdminSuccessResponse:
      description: Standard successful admin response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AdminStandardResponse'
...
    AdminBadRequestResponse:
      description: Standard admin bad request response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AdminStandardResponse'
    AdminForbiddenResponse:
      description: Standard admin forbidden response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AdminStandardResponse'
    AdminNotFoundResponse:
      description: Standard admin not found response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AdminStandardResponse'
    AdminInternalErrorResponse:
      description: Standard admin internal error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AdminStandardResponse'
  schemas:
...
    AdminMeta:
      type: object
      description: Standard pagination metadata envelope for admin responses.
      properties:
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        total_pages:
          type: integer
          example: 5
    AdminStandardResponse:
      type: object
      description: Standard BFF admin response shape.
      properties:
        success:
          type: boolean
          example: true
        data:
          nullable: true
          type: object
        error:
          type: string
          example: ''
        meta:
          $ref: '#/components/schemas/AdminMeta'

    # ── Auth ───────────────────────────────────────────────────────────────────
...